homework-agents/requirements.txt
langchain
langchain-community
fastmcp
ollama
python-dotenv
